auth-service/requirements.txt
celery==5.3.4
circuitbreaker==2.0.0
Django==4.2.5
django-cors-headers==4.2.0
django-filter==23.3
djangorestframework==3.14.0
flake8==6.1.0
flake8-html==0.4.3
flower==2.0.1
Pillow==10.1.01
PyJWT==2.8.0
pytest==7.4.2
pytest-django==4.5.2
redis==5.0.1
requests==2.31.0
salvo==0.2